requirement.txt
requests
urllib
pandas 
openpyxl
langchain-community
langchain-core
langchain-chroma
langchain-ollama
chroma
numpy
uuid
tiktoken
openmeteo-requests
requests-cache 
retry-requests
google-generativeai
bs4
langgraph>=0.0.15
langchain>=0.0.350
langchain-google-genai>=0.0.6
sqlite3-api>=2.0.1
